fix(admin): payment 退款金额 Long（分）+ ghost 字段清除——出口删 auditorId/auditedAt（issue #59，#55 T3）
对照兄弟仓 payment 真实契约：RefundOrderResponse.refundAmount 为 Long、从不发送
auditorId/auditedAt（审核人只出 auditorName，时间维度是 approvedAt/refundedAt 等）；
RefundOrderQuery 有 auditorId 筛选、金额区间为 Long。admin 侧退款上下文单方面对齐
（ADR-0011 §1/§3），payment 零改动：

- 退款列表 + 详情四个 DTO（RefundOrderWireResponse / RefundOrderDetailWireResponse /
  RefundOrderSummaryResponse / RefundOrderDetailResponse）refundAmount
  BigDecimal → Long；删 ghost 字段 auditorId/auditedAt（auditorName 保留——
  provider 实有）
- 北向 RefundOrderQuery 与 wire RefundOrderListWireRequest 的 refundAmountMin/Max
  同步 Long；javadoc 撤销「暂以元承」立场、更正为分（ADR-0011）
- 北向退款 Query 的 auditorId 筛选保留（provider 支持按审核人筛选，照常透传）
- 零加戏：provider 实有而未暴露的字段（refundableAmount/approvedAt/auditAgreed/
  auditRemark 等）一个不补（ADR-0011 §4，展现需求联调后从 UI 起另提）
- 枚举 *Name 透传契约测试退款 fixture 改 provider 真实形状：金额 string 形态
  （"refundAmount": "9900"——框架 Long→ToStringSerializer）+ ghost 字段移除
  + 新增 Long 透传断言（string 进 Long 出，零换算）
- BFF 集成测试退款 fixture 同步 Long 分（99.00→9900L 等）、构造器 arity 收窄、
  删 ghost 断言；RBAC 集成测试 null fixture arity 13→11

北向 JSON 退款金额随之变 string 形态；前端联动见 aieducenter-admin-web#54
变更点 1（金额 string）+ 变更点 4（退款列表/详情删两字段、auditorId 筛选保留）。

stats 金额 + refundBacklog 嵌套（#60）另票跟进。

+++ src/main/java/com/aieducenter/admin/payment/application/dto/query/RefundOrderQuery.java
@@ -1,6 +1,5 @@
 package com.aieducenter.admin.payment.application.dto.query;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -12,7 +11,8 @@ import java.util.List;
  * 字段以 payment 实现契约为准。</p>
  *
  * <p>枚举筛选项（{@code statuses} / {@code auditType}）以 payment BaseEnum 的 Integer code 传递（payment 按 code 绑定枚举）。
- * 退款金额筛选 {@code refundAmountMin/Max} 暂以元（BigDecimal）承，与 payment 的分（Long）对齐见 #55。</p>
+ * 退款金额筛选 {@code refundAmountMin/Max} 为 <strong>Long（分）</strong>，与 payment {@code RefundOrderQuery}
+ * 同型透传、零换算（ADR-0011——前端已按分提交，admin 不做分↔元换算）。</p>
  *
  * @since 0.1.0
  */
@@ -36,14 +36,14 @@ public record RefundOrderQuery(
         /** 审核类型（payment BaseEnum code：AUTO 免审 / MANUAL 人工） */
         Integer auditType,
 
-        /** 审核人 ID（operator id） */
+        /** 审核人 ID（operator id；payment {@code RefundOrderQuery} 支持按审核人筛选） */
         Long auditorId,
 
-        /** 退款金额下限（含） */
-        BigDecimal refundAmountMin,
+        /** 退款金额下限（分，含） */
+        Long refundAmountMin,
 
-        /** 退款金额上限（含） */
-        BigDecimal refundAmountMax,
+        /** 退款金额上限（分，含） */
+        Long refundAmountMax,
 
         /** 创建时间起（含） */
         LocalDateTime createdAtFrom,
